+++ packages/webcomponents/src/components/checkout/save-new-payment-method.tsx
@@ -1,27 +1,32 @@
-import { Component, h, State, Event, EventEmitter } from '@stencil/core';
+import { Component, h, State, Event, EventEmitter, Prop } from '@stencil/core';
+import { StyledHost } from '../../ui-components';
+import { checkoutStore } from '../../store/checkout.store';
 
 @Component({
   tag: 'justifi-save-new-payment-method',
+  shadow: true
 })
 export class SaveNewPaymentMethod {
+  @Prop() label?: string = 'Save New Payment Method';
   @State() isChecked: boolean = false;
   @Event({ bubbles: true }) checkboxChanged: EventEmitter<boolean>;
 
-  handleCheckboxChange(_name, value) {
+  private handleCheckboxChange = (_name: string, value: boolean) => {
     this.isChecked = value;
+    checkoutStore.savePaymentMethod = value;
     this.checkboxChanged.emit(this.isChecked);
   }
 
   render() {
     return (
-      <div class="mt-4 form-check">
+      <StyledHost>
         <form-control-checkbox
-          label="Save new payment method"
+          label={this.label || 'Save New Payment Method'}
           name="saveNewPaymentMethod"
           checked={this.isChecked}
-          inputHandler={this.handleCheckboxChange.bind(this)}
+          inputHandler={this.handleCheckboxChange}
         />
-      </div>
+      </StyledHost>
     );
   }
 }