+++ packages/webcomponents/src/index.ts
@@ -1,1 +1,8 @@
 export * from './components';
+export { PAYMENT_METHODS } from './components/modular-checkout/ModularCheckout';
+export type { CheckoutChangedEventDetail } from './components/modular-checkout/ModularCheckout';
+export type { ICheckoutPaymentMethod } from './api';
+export type {
+  SelectedPaymentMethod,
+  SavedPaymentMethod,
+} from './components/modular-checkout/ModularCheckout';